Plugin: Introduce `tools` folder with configuration files
