+++ packages/react-native-editor/.eslintrc.js
@@ -15,7 +15,9 @@ module.exports = {
 			version: 'detect',
 			flowVersion: '0.92.0',
 		},
-		'import/resolver': require.resolve( '../../test/lint/import-resolver' ),
+		'import/resolver': require.resolve(
+			'../../tools/eslint/import-resolver'
+		),
 	},
 	rules: {
 		'no-restricted-syntax': [
